move most code to config module
